Finish day 9.

+++ src/2019/q05.rs
@@ -5,13 +5,16 @@ use crate::intcode::{Intcode, State};
 
 static INPUT: &str = include_str!("data/q05.data");
 
-fn process_data_a(data: &str) -> i64 {
-    let ints: Vec<i64> = data.split(',').map(|i| i.parse::<i64>().unwrap()).collect();
+fn process_data_a(data: &str) -> i128 {
+    let ints: Vec<i128> = data
+        .split(',')
+        .map(|i| i.parse::<i128>().unwrap())
+        .collect();
     let mut machine = Intcode::new(ints, vec![1]);
     match machine.run_tape() {
         Ok(State::Halted) => {
             let final_value = *machine.outputs.last().unwrap();
-            if machine.outputs.iter().sum::<i64>() == final_value {
+            if machine.outputs.iter().sum::<i128>() == final_value {
                 return final_value;
             }
         }
@@ -25,8 +28,11 @@ fn process_data_a(data: &str) -> i64 {
     -1
 }
 
-fn process_data_b(data: &str) -> i64 {
-    let ints: Vec<i64> = data.split(',').map(|i| i.parse::<i64>().unwrap()).collect();
+fn process_data_b(data: &str) -> i128 {
+    let ints: Vec<i128> = data
+        .split(',')
+        .map(|i| i.parse::<i128>().unwrap())
+        .collect();
     let mut machine = Intcode::new(ints, vec![5]);
     match machine.run_tape() {
         Ok(State::Halted) => {
